requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
scanpy
scvi-tools
scarches
torch
leidenalg
igraph
